added CommandRunner class
