+++ src/Game.py
@@ -1,12 +1,13 @@
 from src.GameState import GameState
 from src.InputHandle import InputHandle
+from src.CommandRunner import CommandRunner
 
 class Game:
 
     def __init__(self, map: GameState):
         self.game_state = map
-        self.input_handle = InputHandle(map)
-        ...
+        self.input_handler = InputHandle()
+        self.command_runner = CommandRunner(self.game_state)
 
     def run(self):
         # mainloop
@@ -19,6 +20,5 @@ class Game:
             if inp == "q":
                 return
 
-            commands_to_run = self.input_handle.parse_user_input(inp)
-
-            self.input_handle.run_commands(commands_to_run)
+            commands_to_run = self.input_handler.parse_user_input(inp)
+            self.command_runner.execute(commands_to_run)
